Added Menu
I only made some small changes in the main for the MainComponent and the Menu as it is now is rather rudimentary since I am still not sure of a design. The menu itself is a frame, so we can add or change it in any way we want. The help file needs to be in the projects main directory and Menu in a folder menu folder under cinnamon.